extract mocks into separate file. fix readbyte mock return value to match byte array length
